Fixed linearInterpolationQuaternion() to avoid returning NaN values in case input quaternions are the same.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/quartzcore/trunk@35715 72102866-910b-0410-8b05-ffd578937521

+++ Source/CATransform3D.m
@@ -242,7 +242,7 @@ CATransform3D CATransform3DInvert(CATransform3D t)
 /* code for debug output:
   for (int i = 0; i < 16; i++)
     {
-      printf("%g ", ((GLfloat*)&transform)[i]);
+      printf("%g ", ((CGFloat*)&transform)[i]);
       if (i%4 == 3) 
 	printf("\n");
     }
